requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
black==24.8.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-filter==24.3
django-prometheus==2.3.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
rpds-py==0.20.0
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
